added H2db functionalities

+++ src/main/java/com/dev/todoApplication/todo/TodoController.java
@@ -56,7 +56,7 @@ public class TodoController {
     }
 
     @RequestMapping(value="/update-todo", method = RequestMethod.POST)
-    public String updateTodo(@RequestParam long id, @RequestParam String updateField,@RequestParam String updatedValue, ModelMap model){
+    public String updateTodo(@RequestParam int id, @RequestParam String updateField,@RequestParam String updatedValue, ModelMap model){
         if(updateField.equals("Choose Field")|| updatedValue==""){
             model.put("Error","Enter valid fields to update");
             return "updateList";
